WIP- Added a StatsdUDP partial class for syncronous methods.  Unit tests pending
